Split query-introspection-enum into its own file
